Rename LICENSE to LICENSE.txt
This change is minor, but is done so that the file has an appropriate
extension when viewed from file browsers like Explorer on Windows,
which can associated '.txt' to text files (rather than not having
an associated handler for an unsuffixed filetype).

From GitHub's perspective, this change is unnecessary, as it only looks
for 'LICENSE', but does not care if it is without an extension, or with
an 'md'/'txt' extension.